+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
+- idle timeouts firing too early after idle duration is over
 - undefined behaviour when idle mode is interrupted from the `onBegin` callback
 
 ## [0.2.1] - 2021-05-27